Quick fix for RabbitMQ/logstash.
For some reason logstash wasn't connecting up the queue correctly to the
exchange that resulted in no messages getting put into the queue and so
nothing was ending up in the index. I don't really understand why this
was working before but don't have time to dig into it now.